--- .eslintrc.json
@@ -1,45 +1,0 @@
-{
-  "env": {
-    "browser": true,
-    "commonjs": true,
-    "es6": true,
-    "node": true
-  },
-  "extends": "airbnb-base",
-  "globals": {
-    "Atomics": "readonly",
-    "SharedArrayBuffer": "readonly"
-  },
-  "parserOptions": {
-    "ecmaVersion": 2018
-  },
-  "rules": {
-  
-
-    "camelcase": 2,
-
-    "eqeqeq":2,
-
-    "no-restricted-globals":0,
-
-    "indent": 2,
-
-
-
-    "no-console" : 0,
-
-    "prefer-destructuring" : 0,
-
-    "no-nested-ternary" : 0,
-
-    "no-unused-expressions" : 0,
-
-    "implicit-arrow-linebreak" : 0,
-
-    "comma-dangle" : 0,
-
-    "no-lonely-if" : 0,
-
-    "no-shadow" : 0
-  }
-}
