+++ src/controllers/index.js
@@ -1,7 +1,6 @@
 const express = require("express");
 const signup = require("./signup");
-//import {get} as getLogin from "./userControllers/login" ;
-const {get: getLogin} = require('./userControllers/login')
+const login = require('./login')
 const {serverError,clientError} = require('./error');
 const userControllers = require("./userControllers")
 
@@ -10,11 +9,13 @@ const router = express.Router();
 //no need for auth
 router.get("/signup", signup.get);
 router.post("/signup", signup.post);
-router.get("/", getLogin);
+router.post("/login", login.post);
+router.get("/login", login.get);
+
 //need auth middleware
 router.use(userControllers);//user middleware
 
-router.use(clientError);
+//router.use(clientError);  No need cause all routes serve home
 router.use(serverError);
