+++ src/controllers/userControllers/like.js
@@ -8,13 +8,3 @@ exports.post = (req, res) => {
     res.redirect('/posts');
   });
 };
-
-// exports.post = (req, res) => {
-//     let id = req.body.id;
-//     let title = req.body.title;
-//     let desc = req.body.desc;
-//     queries.addNewPost(id, title, desc, likes, (err, result) => {
-//         if (err) console.log("ERROROROR");
-//         res.redirect('/posts');
-//     })
-// }
