+++ src/controllers/userControllers/posts.js
@@ -6,7 +6,7 @@ exports.get = (req, res) => {
   queries.selectAllFrom("blogs", (err, result) => {
     if (err) console.log("ERRROR");
 
-    let decoded = decode(ppcookie.parse(req.headers.cookie).udetails);
+    let decoded = decode(ppcookie.parse(req.headers.cookie).uDetails);
 
     queries.selectUserByUserName(decoded.uUser, (err, resu) => {
       if (err) console.log(err);
@@ -33,7 +33,7 @@ exports.addPostGet = (req, res) => {
 };
 
 exports.addPost = (req, res) => {
-  let decoded = decode(ppcookie.parse(req.headers.cookie).udetails);
+  let decoded = decode(ppcookie.parse(req.headers.cookie).uDetails);
   queries.selectUserByUserName(decoded.uUser, (err, resu) => {
     if (err) console.log(err);
     let user = resu.rows[0];
@@ -59,7 +59,6 @@ exports.editPost = (req, res) => {
 exports.deleteBlog = (req, res) => {
   queries.deleteByIdFrom(req.body.id, "blogs", (err, result) => {
     if (err) console.log(err);
-    if (result) console.log(result);
     res.redirect("/posts");
   });
 };
